Fixed the lazyloader contract + tests: no more Nones
There was a weird bug.
When N+1 bulk lazy loading was triggered, the value was loaded alright,
and then immediately reset to `None`. It was really hard to catch.
And it turned out that in the end, it wasn't loaded at all :)

Then I realized that my unit-tests didn't really check if the value was
loaded correctly; they only tested that it was loaded.

In the end I discovered that my lazy loader has a row processing
function that was supposed to *return the loaded value*. But in fact, it
never did. A `None` was returned by default, and set on every instance.

So in the end, lazy loading worked fine, and then was immediately reset.

This fix honors the contract and actually returns the loaded value.