fix error when lazy-loading MappedCollection relationships
When you try to apply nplus1loader to some mapped_collection(),
you get an exception on attribute access:

    AttributeError: 'str' object has no attribute '...'

That happens because SQLAlchemy always populates collections
from a list, even if the collection is dict-like.

But, nplus1loader gives it a dict instead of list.

Then SQLAlchemy tries to walk across it using `for ... in` loop,
and with `dict` that means that it walks over keys (while really
it should walk over ORM instances).

So it tries to rach some attribute of the string key,
and fails with the exception listed above.

The solution is hacky: discard dict keys, and feed only values
(the list of ORM instances) back to the SQLAlchemy.

That is slightly inoptimal (since we're poplating each
MappedCollection exaclty twice, and the first copy is thrown away),
but at least, it works.